Started refactoring for forward-compatible merge support

+++ shared/services/wallet/fees/lighthouse/fee-recipient-manager.go
@@ -4,13 +4,10 @@ import (
 	"fmt"
 	"io/fs"
 	"io/ioutil"
+	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/rocket-pool/rocketpool-go/minipool"
-	"github.com/rocket-pool/rocketpool-go/node"
-	"github.com/rocket-pool/rocketpool-go/rocketpool"
 	"github.com/rocket-pool/smartnode/shared/services/config"
 	"github.com/rocket-pool/smartnode/shared/services/wallet/keystore"
 )
@@ -31,44 +28,48 @@ func NewFeeRecipientManager(keystore keystore.Keystore) *FeeRecipientManager {
 	}
 }
 
-// Creates a fee recipient file that points all of this node's validators to the node distributor address.
-func (fm *FeeRecipientManager) StoreFeeRecipientFile(rp *rocketpool.RocketPool, nodeAddress common.Address) (common.Address, error) {
+// Checks if the fee recipient file exists and has the correct distributor address in it.
+// If it does, this returns true - the file is up to date.
+// Otherwise, this writes the file and returns false indicating that the VC should be restarted to pick up the new file.
+func (fm *FeeRecipientManager) CheckAndUpdateFeeRecipientFile(distributor common.Address) (bool, error) {
 
-	// Get the distributor address for this node
-	distributor, err := node.GetDistributorAddress(rp, nodeAddress, nil)
-	if err != nil {
-		return common.Address{}, fmt.Errorf("error getting distributor address for node [%s]: %w", nodeAddress.Hex(), err)
-	}
-
-	// Write out the default
+	// Create the distributor address string for the node
 	distributorAddress := distributor.Hex()
-	builder := strings.Builder{}
-	builder.WriteString(fmt.Sprintf("default: %s\n", distributorAddress))
+	expectedString := fmt.Sprintf("default: %s\n", distributorAddress)
 
-	// Get all of the minipool addresses for the node
-	minipoolAddresses, err := minipool.GetNodeMinipoolAddresses(rp, nodeAddress, nil)
-	if err != nil {
-		return common.Address{}, fmt.Errorf("error getting minipool addresses for node [%s]: %w", nodeAddress.Hex(), err)
-	}
-
-	// Write all of the validator addresses
-	for _, minipoolAddress := range minipoolAddresses {
-		pubkey, err := minipool.GetMinipoolPubkey(rp, minipoolAddress, nil)
+	// Check if the file exists, and write it if it doesn't
+	path := filepath.Join(fm.keystore.GetKeystoreDir(), config.LighthouseFeeRecipientFilename)
+	_, err := os.Stat(path)
+	if os.IsNotExist(err) {
+		bytes := []byte(expectedString)
+		err = ioutil.WriteFile(path, bytes, FileMode)
 		if err != nil {
-			return common.Address{}, fmt.Errorf("error getting validator pubkey for minipool [%s]: %w", minipoolAddress.Hex(), err)
+			return false, fmt.Errorf("error writing fee recipient file: %w", err)
 		}
 
-		builder.WriteString(fmt.Sprintf("%s: %s\n", pubkey.Hex(), distributorAddress))
+		// If it wrote properly, indicate a success but that the file needed to be updated
+		return false, nil
 	}
 
-	// Write the string out to the file
-	bytes := []byte(builder.String())
-	path := filepath.Join(fm.keystore.GetKeystoreDir(), config.LighthouseFeeRecipientFilename)
-	err = ioutil.WriteFile(path, bytes, FileMode)
+	// Compare the file contents with the expected string
+	bytes, err := ioutil.ReadFile(path)
 	if err != nil {
-		return common.Address{}, fmt.Errorf("error writing fee recipient file: %w", err)
+		return false, fmt.Errorf("error reading fee recipient file: %w", err)
+	}
+	existingString := string(bytes)
+	if existingString != expectedString {
+		// Rewrite the file with the expected distributor address
+		bytes := []byte(expectedString)
+		err = ioutil.WriteFile(path, bytes, FileMode)
+		if err != nil {
+			return false, fmt.Errorf("error writing fee recipient file: %w", err)
+		}
+
+		// If it wrote properly, indicate a success but that the file needed to be updated
+		return false, nil
 	}
 
-	return distributor, nil
+	// The file existed and had the expected address, all set.
+	return true, nil
 
 }
